Kingdom/requirements.txt
altgraph==0.17.3
asgiref==3.7.2
asttokens==2.2.1
auto-py-to-exe==2.25.0
backcall==0.2.0
bottle==0.12.23
bottle-websocket==0.2.9
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==42.0.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-adminlte-ui==2.1.1
django-adminlte3==0.1.6
django-extensions==3.2.1
django-filter==23.5
django-import-export==3.3.1
django-phonenumber-field==7.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Eel==0.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
flake8==6.1.0
future==0.18.2
fuzzywuzzy==0.18.0
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.27.0
greenlet==2.0.1
gunicorn==21.2.0
idna==3.4
import-export==0.3.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.10.0
jedi==0.18.2
Levenshtein==0.20.8
MarkupPy==1.14
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
parso==0.8.3
pefile==2022.5.30
phonenumbers==8.13.20
pickleshare==0.7.5
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.36
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
Pygments==2.14.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==2.8.0
pyparsing==3.0.9
pytest==7.4.3
pytest-django==4.6.0
python-environ==0.4.54
python-Levenshtein==0.20.8
python3-openid==3.2.0
pytz==2023.3.post1
pywin32-ctypes==0.2.0
PyYAML==6.0.1
rapidfuzz==2.13.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.5.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.6
whichcraft==0.6.1
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.5.2
